chore(tui): split cells_components into 7 modules, soften startup colors
Break the monolithic 970-line cells_components.rs into semantic files:
- user_startup.rs     UserCell, StartupCardCell, top_rule_line
- message_cell.rs     MessageCell
- responding_cell.rs  RespondingCell, markdown_body_lines
- thinking_cells.rs   ThinkingTextCell, ThinkingGroupCell
- summary_cells.rs    summary_cell! macro, 5 SummaryCell variants
- plan_cells.rs       PlanSummaryCell, PlanningSuggestionCell, PlanModeCell
- interaction_cells.rs Terminal/Completion/Commit/PendingInteraction cells

Also replace hardcoded Cyan/BOLD/LightBlue colors on the Ready screen
with theme constants, and swap the Startup card border box for a
lightweight centered full-width top-rule line.
