+++ src/tui/render/cells/user_startup.rs
@@ -1,0 +1,95 @@
+use std::path::Path;
+
+use ratatui::{
+    style::{Color, Modifier, Style},
+    text::{Line, Span},
+};
+
+use super::HistoryCell;
+use crate::tui::render::{
+    display_width, prefixed_message_lines, startup_card_inner_width, truncate_for_startup_card,
+    truncate_path_middle, with_border,
+};
+use crate::tui::theme::*;
+
+// ── UserCell ──
+
+pub(crate) struct UserCell {
+    message: String,
+}
+
+impl UserCell {
+    pub(super) fn new(message: impl Into<String>) -> Self {
+        Self {
+            message: message.into(),
+        }
+    }
+}
+
+impl HistoryCell for UserCell {
+    fn display_lines(&self, _width: u16) -> Vec<Line<'static>> {
+        prefixed_message_lines("You", &self.message, 4)
+    }
+}
+
+// ── StartupCardCell ──
+
+pub(crate) struct StartupCardCell {
+    model_label: String,
+    directory: String,
+}
+
+impl StartupCardCell {
+    pub(crate) fn new(model_label: String, directory: String) -> Self {
+        Self {
+            model_label,
+            directory,
+        }
+    }
+}
+
+impl HistoryCell for StartupCardCell {
+    fn display_lines(&self, width: u16) -> Vec<Line<'static>> {
+        let Some(inner_width) = startup_card_inner_width(width) else {
+            return Vec::new();
+        };
+
+        let model_label = "model:";
+        let directory_label = "directory:";
+        let label_width = directory_label.len();
+        let model_prefix = format!("{model_label:<label_width$} ");
+        let hint = "/model to change";
+        let hint_width = display_width(hint);
+        let model_prefix_width = display_width(&model_prefix);
+        let model_available_width = inner_width
+            .saturating_sub(model_prefix_width)
+            .saturating_sub(1)
+            .saturating_sub(hint_width);
+        let model_value = truncate_for_startup_card(&self.model_label, model_available_width);
+        let model_value_width = display_width(&model_value);
+        let gap_width = inner_width
+            .saturating_sub(model_prefix_width)
+            .saturating_sub(model_value_width)
+            .saturating_sub(hint_width)
+            .max(1);
+        let directory_prefix = format!("{directory_label:<label_width$} ");
+        let directory_max_width = inner_width.saturating_sub(display_width(&directory_prefix));
+
+        let lines = vec![
+            Line::from(vec![Span::from(">_ "), Span::from("RARA")]),
+            Line::from(""),
+            Line::from(vec![
+                Span::from(model_prefix),
+                Span::from(model_value),
+                Span::from(" ".repeat(gap_width)),
+                Span::from(hint),
+            ]),
+            Line::from(vec![
+                Span::from(directory_prefix),
+                Span::from(truncate_path_middle(&self.directory, directory_max_width)),
+            ]),
+        ];
+
+        with_border(lines, inner_width)
+    }
+}
